+++ src/qibolab/platforms/abstract.py
@@ -456,7 +456,7 @@ class AbstractPlatform(ABC):
                 from qibolab.platform import Platform
                 from qibolab.sweeper import Sweeper, Parameter
                 from qibolab.pulses import PulseSequence
-                from qibolab.executionparameters import ExecutionParameters
+                from qibolab import ExecutionParameters
 
 
                 platform = Platform("dummy")
