+++ src/qibolab/sweeper.py
@@ -37,7 +37,7 @@ class Sweeper:
             from qibolab.platform import Platform
             from qibolab.sweeper import Sweeper, Parameter
             from qibolab.pulses import PulseSequence
-            from qibolab.executionparameters import ExecutionParameters
+            from qibolab import ExecutionParameters
 
 
             platform = Platform("dummy")
